requirements.txt
djangorestframework
Django==2.0.5
google-api-python-client==1.6.7
httplib2==0.11.3
oauth2client==4.1.2
pyasn1==0.4.3
pyasn1-modules==0.2.1
pytz==2018.4
rsa==3.4.2
six==1.11.0
uritemplate==3.0.0
